fixed crashes when parent script is killed before child script returns fixed crashes when reporting error in an armordesc entry defined in config\itemdesc.cfg fixed warning about streamoff conversion to int

+++ pol-core/pol/module/basicmod.cpp
@@ -465,7 +465,7 @@ namespace Pol {
 	  {
 		ISTRINGSTREAM is( source );
 		string tmp;
-		int tellg;
+		size_t tellg;
 		bool splitted = false;
 
 		while ( is >> tmp )
